+++ src/routes/viewer-paths/+page.server.js
@@ -1,4 +1,4 @@
-import { getCurrentUserSettings, getCurrentUserViewerPaths } from '$lib/server/api/auth_api.js';
+import { getCurrentUserAllowedViewerPaths } from '$lib/server/api/auth_api.js';
 import { getLogger } from '$lib/server/logger.js';
 
 const logger = getLogger('viewer paths page');
@@ -6,11 +6,9 @@ const logger = getLogger('viewer paths page');
 export async function load({ fetch }) {
 	logger.trace('Load viewer paths page');
 
-	const viewerPaths = await getCurrentUserViewerPaths(fetch);
-	const settings = await getCurrentUserSettings(fetch);
+	const viewerPaths = await getCurrentUserAllowedViewerPaths(fetch);
 
 	return {
-		viewerPaths,
-		settings
+		viewerPaths
 	};
 }